[REFACTOR] Aging rules for Education better sintax
