+++ src/test/scala/com/malk/coc/entities/HumanAttributesSpec.scala
@@ -13,7 +13,6 @@ import com.malk.coc.concepts.abstractions.Brain
 
 class HumanAttributesSpec extends AnyFunSpec with Matchers {
   import com.malk.coc.helpers.DiceHelper.implicits._
-  import com.malk.coc.rules.HumanAgingEffectOnEducation.implicits._
   import com.malk.coc.rules.HumanMobility._
   import com.malk.coc.rules.HumanAgingEffectOnAppearance._
