+++ src/test/scala/com/malk/coc/entities/HumanSpec.scala
@@ -36,7 +36,6 @@ class HumanSpec extends AnyFunSpec with Matchers {
     val body = Body(str, con, dex, siz)
     val brain = Brain(int, pow)
 
-    import com.malk.coc.rules.HumanAgingEffectOnEducation.implicits._
     import com.malk.coc.rules.HumanMobility._
     import com.malk.coc.rules.HumanAgingEffectOnAppearance._
